Crate webgrid

This library crate contains all the necessities to run and manage a WebGrid instance.

Submodules have been introduced to split responsibilities. Each module has a specific focus and they together form a chain of dependencies from the low-level library, over the WebGrid domain specific logic, through the executable harness, up to the high-level modules and contained service implementations.

Modules

constants
Constant values

domain
Domain specific structures, implementations, and logic

harness
Runtime harness to execute services in the context of components

library
Independent and project agnostic libraries

module
Runnable modules containing each bundling multiple services and providing a unified configuration

Macros

make_responder_chain_service_fn
Combines a number of responder instances and creates a new hyper service function through the use of make_service_fn. To allow for concurrent access to the responders in the chain, they are wrapped in Arc pointer types.

responder_chain
Chains together a number of Responder implementations