Module webgrid::domain::webdriver

Types in accordance with the WebDriver specification

Structs

Browser
Web browser definition

Capabilities
Struct containing information about the browser requested or provided

CapabilitiesProxy
HTTP proxy settings

CapabilitiesRequest
List of requested capabilities by a client

CapabilityTimeouts
Timeout values for requests to the browser

RawCapabilitiesRequest
Raw CapabilitiesRequest json string containing all fields

ScreenResolution
Resolution of a graphical user interface

SessionCreateResponse
Shell-response to a SessionCreationRequest

SessionCreateResponseValue
Inner response to a SessionCreationRequest

SessionCreationRequest
Incoming request to create a session

WebDriver
Builder for a webdriver instance

WebDriverInstance
Running instance of a WebDriver executable

WebGridOptions
Extension capabilities specific to WebGrid

Enums

BrowserParseError
Error thrown while parsing a browser from string

CapabilityPageLoadStrategy
Describes which DOM event is used to determine whether or not a page has finished loading

CapabilityUnhandledPromptBehavior
How popups like alerts or prompts should be handled

WebDriverError
Errors thrown during creation of a webdriver instance

WebDriverVariant
Vendors for WebDriver executables