Trait webgrid::harness::Service

pub trait Service<F: CommunicationFactory + Send + Sync> {
    type Instance: Send + Sync;
    type Config: Send + Sync;

    const NAME: &'static str;

    fn instantiate(factory: F, config: &Self::Config) -> Self::Instance;
}

Structure which can be instantiated with a CommunicationFactory

Associated Types

type Instance: Send + Sync

Instance type which will be instantiated

type Config: Send + Sync

Configuration type passed to the service

Loading content...
Associated Constants

const NAME: &'static str

Name of the service displayed in log messages

Loading content...
Required methods

fn instantiate(factory: F, config: &Self::Config) -> Self::Instance

Creates a new instance which could be of a different type. This is common when Self: OptionalRequestProcessor where this function would return an instance of Responder containing an instance of Self.

Loading content...