Module webgrid::libraries::helpers

Helper functions that don't belong elsewhere

This module contains small helper functions that don't belong to a specific area but are still used by multiple services.

Modules

constants
Constant values

keys
Redis database keys

lua
Lua functions used for Redis interaction

Structs

Backoff
Exponential backoff iterator

Capabilities
Struct containing information about the browser requested or provided

CapabilitiesProxy
HTTP proxy settings

CapabilitiesRequest
List of requested capabilities by a client

CapabilityTimeouts
Timeout values for requests to the browser

Enums

CapabilityPageLoadStrategy
Describes which DOM event is used to determine whether or not a page has finished loading

CapabilityUnhandledPromptBehavior
How popups like alerts or prompts should be handled

Timeout
Timeout value accessors in seconds

Functions

load_config
Reads a config file by name from the default config directory or one that is specified by the WEBGRID_CONFIG_DIR env variable.

parse_browser_string
Parses a browser string into a name and version

replace_config_variable
Replaces a variable in the passed config string

split_into_two
Splits the input string into two parts at the first occurence of the separator

wait_for
Sends HTTP requests to the specified URL until either a 200 OK response is received or the timeout is reached