Module webgrid::libraries::resources

Monitored resources for jobs

This crate provides resources that are monitoring their internal state. If they become unavailable they report this to a TaskHandle provided by the scheduling module.

Usually this results in the job holding the resource being terminated and restarted.

Structs

DefaultResourceManager
Production resource manager

RedisResource
Redis connection that monitors for connection errors

Enums

PubSubResourceError
PubSub listening errors

Traits

PubSubResource
Redis PubSub channel resource

ResourceManager
Manager that provides access to a set of resources

Type Definitions

PubSub
Boxed PubSubResource shorthand

ResourceManagerResult
Result shorthand

SharedRedisResource
Multiplexed redis resource shared between jobs

StandaloneRedisResource
Individual redis resource created on-demand