Module webgrid::library

Independent and project agnostic libraries

Ideally, any of the library submodules in this module can be extracted into their own crate at any given time (like it happened with jatsl). Libraries in this module have been developed with WebGrid in mind and are powering core functionalities, however, they are in no way bound to the project and everything domain specific has been extracted into the domain module.

Modules

communication
Structures to communication between services in a distributed system

helpers
Various small helper functions

http
Structures for handling and sending HTTP requests

storage
Structures to persist and retrieve arbitrary binary data

Type Definitions

BoxedError
Generic error type

EmptyResult
Result with no value and a BoxedError