Module webgrid::library::communication::discovery

Service discovery and advertisement structures

Modules

pubsub
PubSub channel based discovery implementation

Structs

ServiceAnnouncement
Generic implementation of a service announcement that may be used by implementations of the ServiceAdvertiser and ServiceDiscoverer. Does nothing by itself and is only a shell carrying information.

Traits

DiscoveredServiceEndpoint
Individual endpoint for some service with the optional ability to mark it as unreachable.

ServiceAdvertiser
Structure advertising a service

ServiceDescriptor
Entity which describes a service that can be discovered

ServiceDiscoverer
Structure used to discover endpoints for services

Type Definitions

ServiceEndpoint
Domain specific description of where a service is reachable