Trait webgrid::library::communication::discovery::ServiceAdvertiser

pub trait ServiceAdvertiser {
    #[must_use]
    fn advertise<'life0, 'async_trait, S: ServiceDescriptor + Serialize + Send + Sync, F: Future<Output = ()> + Send + Sync, Fn: FnOnce() -> F + Send + Sync>(
        &'life0 self, 
        service: S, 
        endpoint: ServiceEndpoint, 
        on_ready: Option<Fn>
    ) -> Pin<Box<dyn Future<Output = EmptyResult> + Send + 'async_trait>>
    where
        S: 'async_trait,
        F: 'async_trait,
        Fn: 'async_trait,
        'life0: 'async_trait,
        Self: 'async_trait;
}

Structure advertising a service

Required methods

#[must_use]
fn advertise<'life0, 'async_trait, S: ServiceDescriptor + Serialize + Send + Sync, F: Future<Output = ()> + Send + Sync, Fn: FnOnce() -> F + Send + Sync>(
&'life0 self,
service: S,
endpoint: ServiceEndpoint,
on_ready: Option<Fn>
) -> Pin<Box<dyn Future<Output = EmptyResult> + Send + 'async_trait>> where
S: 'async_trait,
F: 'async_trait,
Fn: 'async_trait,
'life0: 'async_trait,
Self: 'async_trait,

Advertises the given job while the returned future is polled. Additionally, an optional ready function can be passed which is called once the advertiser is ready to accept incoming requests.

Loading content...