Trait webgrid::library::communication::discovery::ServiceDiscoverer

pub trait ServiceDiscoverer<D: ServiceDescriptor> {
    type I: DiscoveredServiceEndpoint;
    fn discover<'a>(
        &self, 
        descriptor: D
    ) -> BoxStream<'a, Result<Self::I, BoxedError>>
    where
        D: 'a;
}

Structure used to discover endpoints for services

Associated Types

type I: DiscoveredServiceEndpoint

Implementation type for the DiscoveredServiceEndpoint

Loading content...
Required methods

fn discover<'a>(
&self,
descriptor: D
) -> BoxStream<'a, Result<Self::I, BoxedError>> where
D: 'a,

Starts the discovery process for a given service. The resulting stream yields DiscoveredServiceEndpoint instances or an error if no more endpoints are available. In case the consumer finds that one of the returned endpoints is unreachable, it should flag it as such so that future discovery attempts will not attempt to re-use the same endpoint (although the specifics are up to the implementation of this trait).

Loading content...