Trait webgrid::library::communication::event::RawNotificationPublisher

pub trait RawNotificationPublisher {
    #[must_use]
    fn publish_raw<'life0, 'life1, 'async_trait>(
        &'life0 self, 
        data: &'life1 [u8], 
        descriptor: QueueDescriptor, 
        extension: Option<QueueDescriptorExtension>
    ) -> Pin<Box<dyn Future<Output = EmptyResult> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: 'async_trait;
}

Structure which allows publishing of serialized data into a queue

Required methods

#[must_use]
fn publish_raw<'life0, 'life1, 'async_trait>(
&'life0 self,
data: &'life1 [u8],
descriptor: QueueDescriptor,
extension: Option<QueueDescriptorExtension>
) -> Pin<Box<dyn Future<Output = EmptyResult> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
Self: 'async_trait,

Sends an opaque payload to a Queue with an optional extension

Loading content...