Module webgrid::library::communication::implementation::json

Serialization and deserialization provided by serde_json using marker traits

This module allows implementors of traits that allow raw access to underlying messaging systems to provide the higher-level traits relying on serialization. It does so by providing a number of marker traits which, when implemented, provide default implementations of the higher-level traits by translating between lower-level serialized data and higher-level strongly typed data by using serde_json. In the future, this will allow for an easy exchange of serialization algorithms by changing the marker traits.

Traits

JsonNotificationPublisher
Marker trait providing a default NotificationPublisher implementation based on serde_json

JsonQueueEntry
Marker trait providing a default QueueEntry implementation based on serde_json

JsonResponseCollector
Marker trait providing a default ResponseCollector implementation based on serde_json

JsonResponsePublisher
Marker trait providing a default ResponsePublisher implementation based on serde_json