Trait webgrid::library::communication::request::Request

pub trait Request: Notification {
    type Response: Serialize + DeserializeOwned + Debug + PartialEq;
    fn reply_to(&self) -> ResponseLocation;
}

Query for information which can be replied to

Note that Requests may not have side effects! Since the response will not be acknowledged, it may get lost and the request can and will get repeated eventually. Thus, in general, there must be no side effects. If, for some sophisticated reason, side effects are required (although they seldom are), they should be idempotent or the response should not be considered important (e.g. it only serves as a confirmation but nothing happens as a consequence of this confirmation — like with a user facing API — in other words a situation where the request will not be repeated if the response is lost).

Associated Types

type Response: Serialize + DeserializeOwned + Debug + PartialEq

Expected response type

Loading content...
Required methods

fn reply_to(&self) -> ResponseLocation

Location where a reply should be sent to

Loading content...