Module webgrid::library::helpers

Various small helper functions

Modules

option_chrono_datetime_as_bson_datetime
Wrapper around chrono_datetime_as_bson_datetime that supports Option<DateTime>

Functions

load_config
Reads a config file by name from the default config directory or one that is specified by the WEBGRID_CONFIG_DIR env variable.

parse_browser_string
Parses a browser string into a name and version

parse_seconds
Parses a Duration from a string containing seconds. Useful for command line parsing

parse_string_list
Parses a string containing a list of strings separated by commas Useful for command line parsing

replace_config_variable
Replaces a variable in the passed config string

split_into_two
Splits the input string into two parts at the first occurence of the separator

wait_for
Sends HTTP requests to the specified URL until either a 200 OK response is received or the timeout is reached