Module webgrid::library::http

Structures for handling and sending HTTP requests

Re-exports

pub use responder::responder_chain;

Structs

MatchableString
Reader-like wrapper for string slices

Enums

ForwardError
HTTP reverse proxy error

Traits

Responder
Handler for incoming HTTP requests which may be chained

Functions

forward_request
Takes an incoming request and forwards it to a remote target

uri_with_authority
Extracts the Uri from a request and replaces the authority with the provided value