Struct webgrid::module::node::Options

pub struct Options {
    pub redis: RedisOptions,
    pub id: Uuid,
    pub initial_timeout: Duration,
    pub idle_timeout: Duration,
    pub webdriver: WebDriverOptions,
    pub recording: ScreenRecordingOptions,
    pub host: String,
    pub bind_timeout: Duration,
    pub storage: StorageOptions,
}

Options for the manager module

Fields
redis: RedisOptionsid: Uuid

Unique identifier of the linked session object

initial_timeout: Duration

Idle timeout (in seconds) which is in effect until the first client request is received. This allows the session to terminate early if the client no longer has any interest in the session or it itself ran into a local timeout (e.g. due to prolonged queueing). After the first request from a client has been received, the regular idle-timeout is taking effect.

idle_timeout: Duration

If no WebDriver client request is received within the specified period, the node will terminate. Each incoming request resets the countdown.

webdriver: WebDriverOptions

Options relating to the WebDriver

recording: ScreenRecordingOptions

Options about screen recording

host: String

Hostname or IP address where this instance can be reached by proxy services

bind_timeout: Duration

Maximum duration (in seconds) for the server to bind to a port and advertise its ready-state. If the server is for whatever reason unable to claim the port in this time, startup will fail.

storage: StorageOptions

Options regarding storage